lookup_plugin: subrules aggregation
when evaluating rules, it is required to split src and dst hosts to guess
it the rule need to be implemented on the target pfsense. Unneccessary hosts
are removed accordingly. In the end, we may generate multiples subrules for
the original rule. This patch generates aliases when required to avoid this
pattern and to generate only one rule in (almost) every case.
Multiple rules are emitted when multiple protocols are used,
since there is no aliases for protocols.