rule: fix tracker id generation
When used with aggregatei module, multiple rules can be generated on the same second,
which cause duplicates IDs. Now, when a rule is created the ID is increment until
an unused one if found. Maybe a check should be also done when the tracker is forced
for an update on an existing rule.