Fix interface log_update when descr is not known
pfSense does not require that the parameter <descr/> exists and is
populated in the XML, especially after first-run. This breaks workflows
which attempt to use Ansible to configure the WAN interface because
lookup of the descr field is absent and execution of the task fails.

In these cases we can fallback to the interface name.